Trait srml_scored_pool::Instance

pub trait Instance: 'static {
    const PREFIX: &'static str;
    const PREFIX_FOR_Pool: &'static str;
    const PREFIX_FOR_CandidateExists: &'static str;
    const PREFIX_FOR_Members: &'static str;
    const PREFIX_FOR_MemberCount: &'static str;
}

Tag a type as an instance of a module.

Defines storage prefixes, they must be unique.

Associated Constants

const PREFIX: &'static str

The prefix used by any storage entry of an instance.

const PREFIX_FOR_Pool: &'static str

const PREFIX_FOR_CandidateExists: &'static str

const PREFIX_FOR_Members: &'static str

const PREFIX_FOR_MemberCount: &'static str

Loading content...